OnClick confirm on button and links

+++ Gestion_des_jeux.php
@@ -92,13 +92,21 @@ if ($stmt = $mysqli->prepare("SELECT nom_jeu, desc_jeu, categorie_jeu, photo_jeu
     echo '<td>'.$row['desc_jeu'].'</td>';
     echo '<td><a href="regle_de_jeu/'.$row['regle_jeu'].'" >Règle de jeu</a></td>';
     echo '<td><a class="btn btn-danger" href="page_modif_jeu.php?modif='.$row['nom_jeu'].'" role="button">Modifier</a></td>';
-    echo '<td><a class="btn btn-danger" href="delete_jeu.php?jeu='.$row['nom_jeu'].'" role="button">Delete</a></td>';
+    echo '<td><a class="btn btn-danger confirmation" href="delete_jeu.php?jeu='.$row['nom_jeu'].'" role="button">Delete</a></td>';
     echo '</tr>';
 $i++;   
 }
 }
 ?>
-
+<script type="text/javascript">
+    var elems = document.getElementsByClassName('confirmation');
+    var confirmIt = function (e) {
+        if (!confirm('Are you sure?')) e.preventDefault();
+    };
+    for (var i = 0, l = elems.length; i < l; i++) {
+        elems[i].addEventListener('click', confirmIt, false);
+    }
+</script>
 </tbody>
 
 </table>
